ndr-api/flowtracex_api/requirements.txt
Django>=4.2
djangorestframework>=3.14
duckdb>=0.9.0
redis>=5.0.0
kafka-python>=2.0.2
psutil>=5.9.0
pandas>=2.0.0
django-cors-headers>=4.0.0
django-filter>=23.0
djangorestframework-simplejwt>=5.3.0
djangorestframework-camel-case>=1.4.0
